requirements.txt
joblib==1.0.1
nltk==3.6.7
numpy==1.21.0
pandas==1.3.0
Pillow==9.1.1
streamlit==1.7.0
tensorflow==2.8.0
